+++ src/Events/EventScriptDeletedEvent.php
@@ -1,0 +1,6 @@
+<?php
+namespace DreamFactory\Core\Script\Events;
+
+class EventScriptDeletedEvent extends BaseEventScriptEvent
+{
+}