+++ src/Events/EventScriptModifiedEvent.php
@@ -1,0 +1,6 @@
+<?php
+namespace DreamFactory\Core\Script\Events;
+
+class EventScriptModifiedEvent extends BaseEventScriptEvent
+{
+}